+++ R/methods-metrics.R
@@ -145,10 +145,13 @@ setMethod(
     signature("SingleCellExperiment"),
     function(object, interestingGroups) {
         validObject(object)
-        colData <- colData(object)
+        if (missing(interestingGroups)) {
+            interestingGroups <- bcbioBase::interestingGroups(object)
+        }
 
+        colData <- colData(object)
+        # Calculate metrics on the fly, if not stashed in colData
         if (!"nUMI" %in% colnames(colData)) {
-            # Calculate on the fly if not stashed
             metrics <- suppressMessages(metrics(
                 object = counts(object),
                 rowData = rowData(object),
@@ -202,10 +205,8 @@ setMethod(
     "metrics",
     signature("seurat"),
     function(object, ...) {
-        fun <- getMethod("metrics", "SingleCellExperiment")
-        data <- fun(object, ...)
-        # Add ident column
-        data[["ident"]] <- slot(object, "ident")
-        data
+        object %>%
+            as("SingleCellExperiment") %>%
+            metrics()
     }
 )