+++ R/methods-plotReadsPerCell.R
@@ -362,7 +362,7 @@ setMethod(
         # Obtain the read counts. Use the unfiltered reads stashed in the
         # metadata if available, otherwise use the metrics return.
         cbList <- metadata(object)[["cellularBarcodes"]]
-        if (is.list(cbList)) {
+        if (length(cbList)) {
             data <- .bindCellularBarcodes(cbList)
         } else {
             data <- metrics(object)